Move spec support into separate files.
