+++ spec/spec_helper.rb
@@ -10,50 +10,8 @@ end
 require "rake-pipeline"
 require "rake-pipeline/filters"
 
-class Rake::Pipeline
-  module SpecHelpers
-
-    # TODO: OS agnostic modules
-    module FileUtils
-      def mkdir_p(dir)
-        system "mkdir", "-p", dir
-      end
-
-      def touch(file)
-        system "touch", file
-      end
-
-      def rm_rf(dir)
-        system "rm", "-rf", dir
-      end
-
-      def touch_p(file)
-        dir = File.dirname(file)
-        mkdir_p dir
-        touch file
-      end
-
-      def age_existing_files
-        old_time = Time.now - 10
-        Dir[File.join(tmp, "**/*.js")].each do |file|
-          File.utime(old_time, old_time, file)
-        end
-      end
-    end
-
-    module Filters
-      ConcatFilter = Rake::Pipeline::ConcatFilter
-
-      class StripAssertsFilter < Rake::Pipeline::Filter
-        def generate_output(inputs, output)
-          inputs.each do |input|
-            output.write input.read.gsub(%r{^\s*assert\(.*\)\s*;?\s*$}m, '')
-          end
-        end
-      end
-    end
-  end
-end
+require "support/spec_helpers/file_utils"
+require "support/spec_helpers/filters"
 
 RSpec.configure do |config|
   original = Dir.pwd
